pyproject.toml
[project]
name = "gerenciador-psicologia"
version = "0.1.0"
description = "Projeto visa criar um sistema para atividades realizadas por um profissional de Psicologia."
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy>=2.0.38",
    "trafilatura>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.3.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["gerenciador_psicologia*"]
exclude = ["static*", "templates*"]
